+++ index.ts
@@ -1,3 +1,4 @@
 export * from './src/components';
 export * from './src/models';
+export * from './src/interfaces';
 export { ColorPickerModule } from './src/color-picker.module';