+++ src/interfaces/color-picker-config.interface.ts
@@ -1,0 +1,6 @@
+export interface IColorPickerConfiguration {
+    width?: number;
+    height?: number;
+    availableColors?: string[];
+    borderRadius?: number;
+}