+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+<a name="WIP"></a>
+# [WIP]() (???)
+
+### Notes
+
+* Refactored the picker component to make use of more syntactic sugar to help tidy up the code.
+* Added documentation to README.md!
+
 <a name="0.2.0"></a>
 # [0.2.0](https://github.com/andymeps/ng2-color-picker/compare/v0.1.3...v0.2.0) (2016-10-03)